+++ frontend/src/Hooks/useEditServices.ts
@@ -1,7 +1,9 @@
 import {ChangeEvent, useState} from "react";
 import {Component} from "../model/Component";
+import {v4 as uuidv4} from "uuid";
+import {ServiceFormDialogProps} from "../Dialog/ServiceFormDialog";
 
-export default function useEditServices() {
+export default function useEditServices(props: ServiceFormDialogProps) {
     const[description, setDescription] =useState<string>("")
     const[workshopName, setWorkshopName] = useState<string>("")
     const[date, setDate] = useState<string>("")
@@ -25,6 +27,28 @@ export default function useEditServices() {
         setWorkshopName("")
         setNewBikeComponents([])
     }
+    function handleSubmitService(){
+        if(props.scrollToBottom){
+            props.scrollToBottom()
+        }
+        props.handleSetServices([...props.editBikeFormState.services,
+            {
+                description: description,
+                newComponents: newBikeComponents,
+                workshopName:workshopName,
+                date: date,
+                id: uuidv4()}])
+        if (props.editMode) {
+            const newComponentCategories = new Set(newBikeComponents
+                .map(component => component.category.trim().toLowerCase()))
+
+            props.handleSetInstalledComponents([...props.editBikeFormState.components.filter(
+                oldComponents => !newComponentCategories.has(oldComponents.category.trim().toLowerCase())),
+                ...newBikeComponents])
+        }
+        clearInputFields()
+        props.handleOpenServiceFormDialog()
+    }
 
     return {
         handleInputDate,
@@ -32,6 +56,7 @@ export default function useEditServices() {
         handleInputWorkshopName,
         clearInputFields,
         handleSetNewComponents,
+        handleSubmitService,
         description,
         workshopName,
         date,
